# swager/openapidist.yaml
openapi: 3.1.0
info:
  version: 1.0.0
  title: Marshrutka
  description: API designed to manage internal procceses of project 'Marshrutka'
servers:
  - url: 'http://localhost:8080'
security:
  - ApiKeyAuth: []
paths:
...
  /customer:
    get:
      operationId: getUser
      description: Retrieves the current authenticated user's information.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User found successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                  user:
                    $ref: '#/components/schemas/CustomerUser'
          links:
            updateUser:
              operationId: saveOrUpdateUser
              description: Update the user profile
            deleteUser:
              operationId: deleteUser
              description: Delete the user account
    post:
      operationId: saveOrUpdateUser
      summary: Save or update user profile
      description: Register or updates the user's information.
      security:
        - bearerAuth: []
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              required:
                - email
                - phoneNumber
                - firstName
                - lastName
                - password
                - dateOfBirth
              properties:
                email:
                  type: string
                  format: email
                firstName:
                  type: string
                  minLength: 1
                  maxLength: 50
                lastName:
                  type: string
                  minLength: 1
                  maxLength: 50
                password:
                  type: string
                  minLength: 6
                  maxLength: 128
                  pattern: '^[A-Za-z\d]{6,}$'
                  description: Password with digit and uppercase letter
                dateOfBirth:
                  type: string
                  format: date
                  example: '1990-01-01'
                phoneNumber:
                  type: string
                  pattern: '^\+?[1-9]\d{1,14}$'
                  example: '+14155552671'
                image:
                  type: string
                  format: binary
                  description: Profile picture file
      responses:
        '200':
          description: User successfully saved
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: user was successfully saved
          links:
            getUser:
              operationId: getUser
              description: Get the updated user info
            deleteUser:
              operationId: deleteUser
              description: Delete the user account
    delete:
      operationId: deleteUser
      summary: Assigns 'Deleted' status to the authenticated user
      security:
        - bearerAuth: []
      responses:
        '200':
          description: User successfully deleted
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: user was successfully deleted
          links:
            createGuest:
              operationId: createGuestUser
              description: Create a guest user after deletion
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
    ApiKeyAuth:
      type: apiKey
      in: header
      name: X-API-Key
  schemas:
...
    CustomerUser:
      type: object
      properties:
        id:
          type: string
          format: uuid
        email:
          type: string
          format: email
        firstName:
          type: string
        lastName:
          type: string
        dateOfBirth:
          type: string
          format: date
        imageUrl:
          type: string
          format: uri
...